Les demandes d'emprunts continuent

+++ database/migrations/2024_05_23_195909_create_soutenances_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             /* $table->string(column : 'year'); */
             $table->date(column : 'start_date');
             $table->date(column : 'end_date');
-            $table->integer(column : 'number_memories_expected');
-            $table->integer(column : 'number_memories_remaining');
+            $table->unsignedBigInteger(column : 'number_memories_expected');
+            $table->unsignedBigInteger(column : 'number_memories_remaining');
             $table->string(column : 'created_by')->nullable()->default(value : NULL);
             $table->string(column : 'updated_by')->nullable()->default(value : NULL);
             $table->string(column : 'deleted_by')->nullable()->default(value : NULL);
